moved some locator related funcitons
